+++ src/NzbDrone.Core/Applications/Readarr/Readarr.cs
@@ -89,36 +89,45 @@ namespace NzbDrone.Core.Applications.Readarr
         {
             _logger.Debug("Updating indexer {0} [{1}]", indexer.Name, indexer.Id);
 
-            var appMappings = _appIndexerMapService.GetMappingsForApp(Definition.Id);
-            var indexerMapping = appMappings.FirstOrDefault(m => m.IndexerId == indexer.Id);
+            var appIndexerProfiles = indexer.AppProfile.FindAll(x => x.Value.ApplicationIDs.Contains(Definition.Id));
 
-            var readarrIndexer = BuildReadarrIndexer(indexer, indexer.Protocol, indexerMapping?.RemoteIndexerId ?? 0);
-
-            var remoteIndexer = _readarrV1Proxy.GetIndexer(indexerMapping.RemoteIndexerId, Settings);
-
-            if (remoteIndexer != null)
+            if (appIndexerProfiles.Count >= 1)
             {
-                _logger.Debug("Remote indexer found, syncing with current settings");
+                var appMappings = _appIndexerMapService.GetMappingsForApp(Definition.Id);
+                var indexerMapping = appMappings.FirstOrDefault(m => m.IndexerId == indexer.Id);
 
-                if (!readarrIndexer.Equals(remoteIndexer))
-                {
-                    _readarrV1Proxy.UpdateIndexer(readarrIndexer, Settings);
-                }
-            }
-            else
-            {
-                _appIndexerMapService.Delete(indexerMapping.Id);
+                var readarrIndexer =
+                    BuildReadarrIndexer(indexer, indexer.Protocol, indexerMapping?.RemoteIndexerId ?? 0);
 
-                if (indexer.Capabilities.Categories.SupportedCategories(Settings.SyncCategories.ToArray()).Any())
+                var remoteIndexer = _readarrV1Proxy.GetIndexer(indexerMapping.RemoteIndexerId, Settings);
+
+                if (remoteIndexer != null)
                 {
-                    _logger.Debug("Remote indexer not found, re-adding {0} to Readarr", indexer.Name);
-                    readarrIndexer.Id = 0;
-                    var newRemoteIndexer = _readarrV1Proxy.AddIndexer(readarrIndexer, Settings);
-                    _appIndexerMapService.Insert(new AppIndexerMap { AppId = Definition.Id, IndexerId = indexer.Id, RemoteIndexerId = newRemoteIndexer.Id });
+                    _logger.Debug("Remote indexer found, syncing with current settings");
+
+                    if (!readarrIndexer.Equals(remoteIndexer))
+                    {
+                        _readarrV1Proxy.UpdateIndexer(readarrIndexer, Settings);
+                    }
                 }
                 else
                 {
-                    _logger.Debug("Remote indexer not found for {0}, skipping re-add to Readarr due to indexer capabilities", indexer.Name);
+                    _appIndexerMapService.Delete(indexerMapping.Id);
+
+                    if (indexer.Capabilities.Categories.SupportedCategories(Settings.SyncCategories.ToArray()).Any())
+                    {
+                        _logger.Debug("Remote indexer not found, re-adding {0} to Readarr", indexer.Name);
+                        readarrIndexer.Id = 0;
+                        var newRemoteIndexer = _readarrV1Proxy.AddIndexer(readarrIndexer, Settings);
+                        _appIndexerMapService.Insert(new AppIndexerMap
+                            {AppId = Definition.Id, IndexerId = indexer.Id, RemoteIndexerId = newRemoteIndexer.Id});
+                    }
+                    else
+                    {
+                        _logger.Debug(
+                            "Remote indexer not found for {0}, skipping re-add to Readarr due to indexer capabilities",
+                            indexer.Name);
+                    }
                 }
             }
         }
@@ -184,9 +193,9 @@ namespace NzbDrone.Core.Applications.Readarr
             {
                 Id = id,
                 Name = $"{indexer.Name} (Prowlarr)",
-                EnableRss = indexer.Enable && indexer.AppProfile.Value.EnableRss,
-                EnableAutomaticSearch = indexer.Enable && indexer.AppProfile.Value.EnableAutomaticSearch,
-                EnableInteractiveSearch = indexer.Enable && indexer.AppProfile.Value.EnableInteractiveSearch,
+                EnableRss = indexer.Enable && enableRss,
+                EnableAutomaticSearch = indexer.Enable && enableAutoSearch,
+                EnableInteractiveSearch = indexer.Enable && enableInteractiveSearch,
                 Priority = indexer.Priority,
                 Implementation = indexer.Protocol == DownloadProtocol.Usenet ? "Newznab" : "Torznab",
                 ConfigContract = schema.ConfigContract,
