+++ src/NzbDrone.Core/Applications/Sonarr/Sonarr.cs
@@ -89,37 +89,44 @@ namespace NzbDrone.Core.Applications.Sonarr
         {
             _logger.Debug("Updating indexer {0} [{1}]", indexer.Name, indexer.Id);
 
-            var appMappings = _appIndexerMapService.GetMappingsForApp(Definition.Id);
-            var indexerMapping = appMappings.FirstOrDefault(m => m.IndexerId == indexer.Id);
+            var appIndexerProfiles = indexer.AppProfile.FindAll(x => x.Value.ApplicationIDs.Contains(Definition.Id));
 
-            var sonarrIndexer = BuildSonarrIndexer(indexer, indexer.Protocol, indexerMapping?.RemoteIndexerId ?? 0);
-
-            var remoteIndexer = _sonarrV3Proxy.GetIndexer(indexerMapping.RemoteIndexerId, Settings);
-
-
-            if (remoteIndexer != null)
+            if (appIndexerProfiles.Count >= 1)
             {
-                _logger.Debug("Remote indexer found, syncing with current settings");
+                var appMappings = _appIndexerMapService.GetMappingsForApp(Definition.Id);
+                var indexerMapping = appMappings.FirstOrDefault(m => m.IndexerId == indexer.Id);
 
-                if (!sonarrIndexer.Equals(remoteIndexer))
-                {
-                    _sonarrV3Proxy.UpdateIndexer(sonarrIndexer, Settings);
-                }
-            }
-            else
-            {
-                _appIndexerMapService.Delete(indexerMapping.Id);
+                var sonarrIndexer = BuildSonarrIndexer(indexer, indexer.Protocol, indexerMapping?.RemoteIndexerId ?? 0);
 
-                if (indexer.Capabilities.Categories.SupportedCategories(Settings.SyncCategories.ToArray()).Any())
+                var remoteIndexer = _sonarrV3Proxy.GetIndexer(indexerMapping.RemoteIndexerId, Settings);
+
+                if (remoteIndexer != null)
                 {
-                    _logger.Debug("Remote indexer not found, re-adding {0} to Sonarr", indexer.Name);
-                    sonarrIndexer.Id = 0;
-                    var newRemoteIndexer = _sonarrV3Proxy.AddIndexer(sonarrIndexer, Settings);
-                    _appIndexerMapService.Insert(new AppIndexerMap { AppId = Definition.Id, IndexerId = indexer.Id, RemoteIndexerId = newRemoteIndexer.Id });
+                    _logger.Debug("Remote indexer found, syncing with current settings");
+
+                    if (!sonarrIndexer.Equals(remoteIndexer))
+                    {
+                        _sonarrV3Proxy.UpdateIndexer(sonarrIndexer, Settings);
+                    }
                 }
                 else
                 {
-                    _logger.Debug("Remote indexer not found for {0}, skipping re-add to Sonarr due to indexer capabilities", indexer.Name);
+                    _appIndexerMapService.Delete(indexerMapping.Id);
+
+                    if (indexer.Capabilities.Categories.SupportedCategories(Settings.SyncCategories.ToArray()).Any())
+                    {
+                        _logger.Debug("Remote indexer not found, re-adding {0} to Sonarr", indexer.Name);
+                        sonarrIndexer.Id = 0;
+                        var newRemoteIndexer = _sonarrV3Proxy.AddIndexer(sonarrIndexer, Settings);
+                        _appIndexerMapService.Insert(new AppIndexerMap
+                            {AppId = Definition.Id, IndexerId = indexer.Id, RemoteIndexerId = newRemoteIndexer.Id});
+                    }
+                    else
+                    {
+                        _logger.Debug(
+                            "Remote indexer not found for {0}, skipping re-add to Sonarr due to indexer capabilities",
+                            indexer.Name);
+                    }
                 }
             }
         }
